Improved: Added fetchGiftCardActivationDetails to OrderService, which was previously an action but did not mutate any state (#516)
