+++ src/store/modules/order/actions.ts
@@ -397,7 +397,7 @@ const actions: ActionTree<OrderState , RootState> ={
     }
 
     if(orderDetails.orderType === 'packed') {
-      order = await dispatch("fetchGiftCardActivationDetails", { isDetailsPage: true, currentOrders: [order] });
+      order = await OrderService.fetchGiftCardActivationDetails({ isDetailsPage: true, currentOrders: [order] });
     }
 
     await dispatch('updateCurrent', { order })
@@ -631,7 +631,7 @@ const actions: ActionTree<OrderState , RootState> ={
         const total = resp.data.grouped?.orderId?.ngroups;
 
         if(payload.viewIndex && payload.viewIndex > 0) orders = state.packed.list.concat(orders)
-        const packedOrders = await dispatch("fetchGiftCardActivationDetails", { isDetailsPage: false, currentOrders: orders })
+        const packedOrders = await OrderService.fetchGiftCardActivationDetails({ isDetailsPage: false, currentOrders: orders })
         commit(types.ORDER_PACKED_UPDATED, { orders: packedOrders, total })
         if (payload.viewIndex === 0) emitter.emit("dismissLoader");
       } else {
@@ -735,73 +735,6 @@ const actions: ActionTree<OrderState , RootState> ={
     }
 
     return resp;
-  },
-
-  async fetchGiftCardActivationDetails({ commit }, { isDetailsPage, currentOrders }) {
-    const orders = JSON.parse(JSON.stringify(currentOrders));
-    const orderIds = [] as any;
-    let giftCardActivations = [] as any;
-
-    if(isDetailsPage) {
-      orderIds.push(orders[0].orderId);
-    } else {
-      orders.map((order: any) => {
-        order.parts.map((part: any) => {
-          part.items.map((currentItem: any) => {
-            if(currentItem.productTypeId === 'GIFT_CARD' && !orderIds.includes(currentItem.orderId)) {
-              orderIds.push(order.orderId);
-            }
-          })
-        })
-      })
-    }
-    if(!orderIds.length) return orders;
-
-    try {
-      const resp = await UtilService.fetchGiftCardFulfillmentInfo({
-        entityName: "GiftCardFulfillment",
-        inputFields: {
-          orderId: orderIds,
-          orderId_op: "in"
-        },
-        fieldList: ["amount", "cardNumber", "fulfillmentDate", "orderId", "orderItemSeqId"],
-        viewSize: 250
-      })
-
-      if(!hasError(resp)) {
-        giftCardActivations = resp.data.docs
-      } else {
-        throw resp.data
-      }
-    } catch(error) {
-      logger.error(error)
-    }
-
-    if(giftCardActivations.length) {
-      if(isDetailsPage) {
-        orders[0].part.items.map((currentItem: any) => {
-          const activationRecord = giftCardActivations.find((card: any) => card.orderId === currentItem.orderId && card.orderItemSeqId === currentItem.orderItemSeqId)
-          if(activationRecord?.cardNumber) {
-            currentItem.isGCActivated = true;
-            currentItem.gcInfo = activationRecord
-          }
-        })
-      } else {
-        orders.map((order: any) => {
-          order.parts.map((part: any) => {
-            part.items.map((item: any) => {
-              const activationRecord = giftCardActivations.find((card: any) => card.orderId === order.orderId && card.orderItemSeqId === item.orderItemSeqId)
-              if(activationRecord?.cardNumber) {
-                item.isGCActivated = true;
-                item.gcInfo = activationRecord
-              }
-            })
-          })
-        })
-      }
-    }
-
-    return isDetailsPage ? orders[0] : orders
   },
 
   async deliverShipment ({ state, dispatch, commit }, order) {
